feat: retrieve active plans
