+++ package.json
@@ -5,6 +5,7 @@
     "dotenv": "^16.4.7",
     "express": "^4.21.1",
     "jsdom": "^25.0.1",
+    "jsonwebtoken": "^9.0.2",
     "node-fetch": "^2.7.0",
     "puppeteer": "^23.9.0",
     "showdown": "^2.1.0",
